third_party/blink/web_tests/fast/css/invalidation/read-only-write-pseudo-expected.txt

Use descendant invalidation sets for :read-only and :read-write pseudo classes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(i1, '').backgroundColor is transparent
PASS getComputedStyle(i2, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(i1, '').backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(i2, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
